student details feature is done
