+++ src/features/index.js
@@ -5,19 +5,6 @@ export * from './auth';
 export * from './routes';
 export * from './header';
 export * from './validation';
-export * from './list-of-groups';
-export * from './list-of-students';
-export * from './list-of-lessons';
-export * from './list-of-courses';
-export * from './list-of-mentors';
-export * from './edit-students-details';
-export * from './edit-group';
-export * from './edit-lesson';
-export * from './edit-mentor';
-export * from './add-lesson';
-export * from './group-details';
-export * from './course-details';
-export * from './student-details';
 export * from './not-found';
 export * from './modal-window';
 export * from './unassigned-list';
@@ -26,4 +13,4 @@ export * from './courses';
 export * from './groups';
 export * from './lessons';
 export * from './secretaries';
-export * from './students'
+export * from './students';
