+++ src/features/students/index.js
@@ -1,3 +1,4 @@
 export * from './edit-students-details';
 export * from './list-of-students';
 export * from './student-profile';
+export * from './student-details';
