+++ src/features/students/student-details/index.js
