+++ src/features/students/student-details/student-details.scss
